feat(pkg/pathfmt): helper for format path add as package
