+++ internal/pkg/pathfmt/pathfmt.go
@@ -1,0 +1,17 @@
+package pathfmt
+
+import "strings"
+
+// RemoveTrailingSlash returns path without trailing slash
+func RemoveTrailingSlash(path string) string {
+	if len(path) > 0 && path[len(path)-1] == '/' {
+		return path[0 : len(path)-1]
+	}
+
+	return path
+}
+
+// RemoveFragment - clear url from reference in path
+func RemoveFragment(path string) string {
+	return RemoveTrailingSlash(strings.Split(path, "#")[0])
+}
